% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf.r
\name{db}
\alias{db}
\title{Separate a table based on two groups of columns.}
\usage{
db(data, ...)
}
\arguments{
\item{data}{}

\item{...}{attributes to separate from the rest as "groups"}
}
\description{
This function will return two tables in a list from one, keeping all columns in one table or the other.
}
\details{
https://en.wikipedia.org/wiki/Database_normalization
}

